Restructure packaging for ideal
COAND-497
